requirement.txt
flask==2.3.2
google-generativeai==0.1.0
requests==2.28.1
pandas==1.4.3  # Example for data manipulation
numpy==1.22.4  # Example for numerical operations
scikit-learn==1.1.1  # Example for machine learning

